# week7/swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Books Management API
  description: |
    API quản lý sách với xác thực JWT.

    **Tài khoản demo:**
    - Username: `admin` / Password: `admin123` (role: admin)
    - Username: `user` / Password: `user123` (role: user)
  contact:
    name: API Support
    email: support@example.com
  version: 1.0.0
servers:
- url: http://localhost:3000/api
  description: Development server
...
paths:
...
  /auth/refresh:
    post:
      tags:
      - authentication
      summary: Cấp lại access token từ refresh token
      description: |
        Nhận access token mới khi access token cũ hết hạn, sử dụng refresh token hợp lệ.
      operationId: refresh_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_refresh_body"
        required: true
      responses:
        "200":
          description: Cấp lại access token thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
        "400":
          description: Yêu cầu không hợp lệ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                message: Title và Author là bắt buộc
        "403":
          description: Token không hợp lệ hoặc đã hết hạn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                message: Token không hợp lệ hoặc đã hết hạn
        "404":
          description: Không tìm thấy tài nguyên
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                message: Không tìm thấy sách
      x-openapi-router-controller: swagger_server.controllers.authentication_controller
  /auth/me:
    get:
      tags:
      - authentication
      summary: Lấy thông tin người dùng hiện tại
      description: Lấy thông tin người dùng đang đăng nhập (yêu cầu token)
      operationId: get_current_user
      responses:
        "200":
          description: Thông tin người dùng
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_2"
        "401":
          description: Chưa xác thực - Token không được cung cấp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                message: Token không được cung cấp
        "403":
          description: Token không hợp lệ hoặc đã hết hạn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                success: false
                message: Token không hợp lệ hoặc đã hết hạn
      security:
      - bearerAuth: []
      x-openapi-router-controller: swagger_server.controllers.authentication_controller
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          description: ID duy nhất của người dùng
          example: 1
        username:
          type: string
          description: Tên đăng nhập
          example: admin
        email:
          type: string
          description: Email người dùng
          format: email
          example: admin@example.com
        role:
          type: string
          description: Vai trò của người dùng
          example: admin
          enum:
          - user
          - admin
      example:
        role: admin
        id: 1
        email: admin@example.com
        username: admin
...
    Error:
      required:
      - message
      - success
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
...
    auth_refresh_body:
      required:
      - refresh_token
      type: object
      properties:
        refresh_token:
          type: string
          description: Refresh token hợp lệ
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    inline_response_200_1:
      type: object
      properties:
        success:
          type: boolean
          example: true
        access_token:
          type: string
          description: Access token mới
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      example:
        access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        success: true
    inline_response_200_2:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: "#/components/schemas/User"
      example:
        data:
          role: admin
          id: 1
          email: admin@example.com
          username: admin
        success: true
...
  securitySchemes:
    bearerAuth:
      type: http
      description: |
        Nhập JWT token nhận được từ endpoint `/auth/login` hoặc `/auth/register`.

        Format: `Bearer YOUR_JWT_TOKEN`
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth